backend_api/requirements.txt
asgiref==3.8.1
Django==3.2
django-cors-headers==4.5.0
djangorestframework==3.12.0
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.13.2
whitenoise==6.9.0
